# public/v1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "ai",
    "version": "1.0.0",
    "title": "ai",
    "termsOfService": "http://pandariel.com",
    "contact": {
      "email": "huayun321@gmail.com"
    }
  },
  "host": "immense-lowlands-91960.herokuapp.com",
  "basePath": "/v1",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/client/article/create": {
      "post": {
        "tags": [
          "article",
          "client"
        ],
        "summary": "创建文章",
        "description": "客户端创建文章，需要用户权限，即传token",
        "operationId": "CreateArticle",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "参数",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ArticleCreateForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/SuccessModel"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          },
          "500": {
            "description": "internal error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
...
    "/admin/article/select": {
      "post": {
        "tags": [
          "article",
          "admin"
        ],
        "summary": "加精选",
        "description": "加精选",
        "operationId": "SelectArticle",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "参数",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IDForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/SuccessModel"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          },
          "500": {
            "description": "internal error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/admin/article/un-select": {
      "post": {
        "tags": [
          "article",
          "admin"
        ],
        "summary": "取消精选",
        "description": "取消精选",
        "operationId": "UnSelectArticle",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "参数",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IDForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/SuccessModel"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          },
          "500": {
            "description": "internal error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/admin/article/delete": {
      "post": {
        "tags": [
          "article",
          "admin"
        ],
        "summary": "删除",
        "description": "删除",
        "operationId": "DeleteArticle",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "参数",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IDForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/SuccessModel"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          },
          "500": {
            "description": "internal error",
            "schema": {
              "$ref": "#/definitions/ErrorModel"
            }
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "auth": {
      "type": "apiKey",
      "name": "token",
      "in": "query"
    },
    "auth_header": {
      "type": "apiKey",
      "name": "Authorization",
      "description": "value should be: bearer {token}",
      "in": "header"
    }
  },
  "definitions": {
...
    "IDForm": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
...
    "ArticleCreateForm": {
      "type": "object",
      "required": [
        "author_id",
        "content"
      ],
      "properties": {
        "author_id": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "ErrorModel": {
      "type": "object",
      "required": [
        "message",
        "code"
      ],
      "properties": {
        "message": {
          "type": "string"
        },
        "code": {
          "type": "integer"
        }
      }
    },
    "SuccessModel": {
      "type": "object",
      "required": [
        "message",
        "code"
      ],
      "properties": {
        "message": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "minimum": 0,
          "maximum": 20000
        }
      }
    },
...
  }
}